move getCurrentClient
