+++ src/navigation/RootNavigator.tsx
@@ -8,8 +8,9 @@ import MiscStack from 'navigation/MiscStack';
 import FormsStack from 'navigation/FormsStack';
 import firebase from 'database/clientApp';
 import { ClientContext } from 'context/ContextProvider';
-import { getCurrentClient } from 'database/auth';
 import { getEmptyClient } from 'utils/utils';
+import { getClient } from 'database/queries';
+import { Client } from 'types/types';
 
 const auth = firebase.auth();
 const Stack = createStackNavigator();
@@ -22,6 +23,15 @@ export const styles = StyleSheet.create({
   },
 });
 
+async function getCurrentClient(): Promise<Client> {
+  const uid = firebase.auth().currentUser?.uid;
+  if (uid !== undefined) {
+    const client = await getClient(uid);
+    return client;
+  }
+  throw new Error('could not fetch current client from firebase');
+}
+
 export default function RootNavigator() {
   const [user, setUser] = useState<firebase.User | null>(null);
   const [isLoading, setIsLoading] = useState(true);
@@ -31,12 +41,10 @@ export default function RootNavigator() {
     const unsubscribeAuth = auth.onAuthStateChanged(async authenticatedUser => {
       try {
         if (authenticatedUser) {
-          console.log('logged in');
           await setUser(authenticatedUser);
           const client = await getCurrentClient();
           update(client); // update app context
         } else {
-          console.log('logged out');
           await setUser(null);
           update(getEmptyClient()); // update app context
         }
